requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.2.0
distlib==0.3.9
filelock==3.16.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
googlesearch-python==1.3.0
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
propcache==0.3.1
PyJWT==2.10.1
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
twilio==9.6.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.29.1
Werkzeug==3.1.3
yarl==1.20.0
